fix: commonises remove applicants function
